update the lat and long precisions

+++ _6_Coordinate_EP/k7080-HigherResolution/0.1_child.c
@@ -294,9 +294,9 @@ int closetGridIndex(float bldlat, float bldlong){
 void receiveLongLat(void) {
     // MPI_Recv(&msg_arr, 3, MPI_FLOAT, MPI_ANY_SOURCE, MPI_ANY_TAG, parent_comm, &status);
     // MPI_Send(&data, performanc_length, MPI_FLOAT,status.MPI_SOURCE, 0, parent_comm);
-    MPI_Recv(&latall, INNERMOST_POINTS * INNERMOST_POINTS, MPI_FLOAT, 
+    MPI_Recv(&latall, INNERMOST_POINTS * INNERMOST_POINTS, MPI_DOUBLE, 
         MPI_ANY_SOURCE, MPI_ANY_TAG, parent_comm, &status);
-    MPI_Recv(&longall, INNERMOST_POINTS * INNERMOST_POINTS, MPI_FLOAT,
+    MPI_Recv(&longall, INNERMOST_POINTS * INNERMOST_POINTS, MPI_DOUBLE,
         MPI_ANY_SOURCE, MPI_ANY_TAG, parent_comm, &status);
     // print the received latlongalls
     for (int k = 0; k < INNERMOST_POINTS * INNERMOST_POINTS; k++) {
